requirements.txt
jupyter
numpy
pandas
locust
qdrant_client
pinecone-client